Use scip-java snapshot logic
Previously, this repo had custom logic to snapshot the generated
SemanticDB files, which was missing some SCIP features like
relationships. This commit replaces the custom snapshot logic with the
`scip-java snapshot` command, which prints all SCIP features including
symbol relationships.
