--- semanticdb-kotlinc/minimized/src/generatedSnapshots/resources/java/snapshots/CompanionConsumer.java
@@ -1,15 +1,0 @@
-package snapshots;
-
-public class CompanionConsumer {
-//           ^^^^^^^^^^^^^^^^^ definition snapshots/CompanionConsumer# public class CompanionConsumer
-    CompanionConsumer() {
-//  ^^^^^^^^^^^^^^^^^ definition snapshots/CompanionConsumer#`<init>`(). CompanionConsumer()
-        CompanionOwner.Companion.create();
-//      ^^^^^^^^^^^^^^ reference snapshots/CompanionOwner#
-//                     ^^^^^^^^^ reference snapshots/CompanionOwner#Companion.
-//                               ^^^^^^ reference snapshots/CompanionOwner#Companion#create().
-        new CompanionOwner().create();
-//          ^^^^^^^^^^^^^^ reference snapshots/CompanionOwner#`<init>`().
-//                           ^^^^^^ reference snapshots/CompanionOwner#create().
-    }
-}
